fix: add discount for order

+++ backend/src/main/java/secure_shop/backend/dto/discount/DiscountDetailsDTO.java
@@ -20,7 +20,7 @@ public class DiscountDetailsDTO {
     private DiscountType discountType;
     private BigDecimal discountValue;
     private BigDecimal minOrderValue;
-    private BigDecimal maxUsage;
+    private Integer maxUsage;
     private Integer perUserLimit;
     private Integer used;
     private Instant startAt;
